improving code
