+++ src/game_engine/render_scene.c
@@ -6,38 +6,11 @@
 /*   By: gmasid <gmasid@student.42.rio>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/05 18:11:07 by gmasid            #+#    #+#             */
-/*   Updated: 2023/04/09 12:46:54 by gmasid           ###   ########.fr       */
+/*   Updated: 2023/04/09 15:53:53 by gmasid           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
-
-void	set_pixel_color(int x, int y, int color, t_game *game)
-{
-	char	*dst;
-	t_img	*img;
-
-	img = &game->window_image;
-	dst = img->addr + (y * img->line_len + x * (img->bpp / 8));
-	*(unsigned int *)dst = color;
-}
-
-int	is_within_ray_limit(int x, int y, t_game *game)
-{
-	float	delta_x;
-	float	delta_y;
-	float	distance;
-
-	delta_x = x - game->player.x_pos - 0.5;
-	delta_y = y - game->player.y_pos - 0.5;
-	distance = sqrt(powf(delta_x, 2.) + powf(delta_y, 2.));
-	return (distance < game->ray.limit);
-}
-
-int	is_wall(int x, int y, t_game *game)
-{
-	return (game->map.array[x][y] == '1');
-}
 
 void	render_wall_column_pixel(int x, int y, int wall_height, t_game *game)
 {
